% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_array3d}
\alias{fill_array3d}
\title{fill_array3d}
\usage{
fill_array3d(mat, newmat, i_start = 1L, j_start = 1L, k_start = 1L)
}
\arguments{
\item{mat}{A 3d Array.}

\item{newmat}{A 3d Array to fill \code{mat}.}

\item{i_start}{Integer, row of \code{mat} starting to fill.}

\item{j_start}{Integer, col of \code{mat} starting to fill.}

\item{k_start}{Integer, slice of \code{mat} starting to fill.}
}
\description{
This function is just for Rcmip5 package when combine 3d array.
}
\details{
Note that if data types of \code{mat} and \code{newmat} are different, 
values in \code{mat} will be \strong{not replaced.}
}
\examples{
mat    <- array(2, dim = c(3, 3, 3))
newmat <- array(1, dim = c(3, 3, 3))
fill_array3d(mat, newmat)
mat

}
\references{
[1] https://stackoverflow.com/questions/30969799/column-means-3d-matrix-cube-rcpp \cr
[2] https://stackoverflow.com/questions/34161593/how-to-replace-elements-of-a-matrix-in-c-with-values-from-another-matrix-usin \cr
}
