% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_matrix}
\alias{fill_matrix}
\alias{fill_matrix_rcpp}
\alias{fill_matrix_arma}
\title{fill_matrix

Replace sub matrix in rcpp format.}
\usage{
fill_matrix_rcpp(mat, newmat, i_start = 1L, j_start = 1L)

fill_matrix_arma(mat, newmat, i_start = 1L, j_start = 1L)
}
\arguments{
\item{mat}{A matrix.}

\item{newmat}{A matrix to fill \code{mat}.}

\item{i_start}{Integer, row of \code{mat} starting to fill.}

\item{j_start}{Integer, col of \code{mat} starting to fill.}
}
\description{
fill_matrix

Replace sub matrix in rcpp format.
}
\examples{
# mat   : ones matrix
# newmat: 1:16 matrix

## 1. rcpp version
mat <- matrix(1, 6, 6)
newmat <- matrix(1:16, 4, 4)*1.0
fill_matrix_rcpp(mat, newmat)
mat

## 2. RcppArmadillo version
mat <- matrix(1, 6, 6)
newmat <- matrix(1:16, 4, 4)
fill_matrix_arma(mat, newmat)
mat
}
