% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppArray-package.R
\docType{package}
\name{RcppArray-package}
\alias{RcppArray}
\alias{RcppArray-package}
\title{RcppArray: Avoid Deep Copy in Matrix and Array.}
\description{
Avoid Deep Copy in Matrix and Array.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
