% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{test}
\alias{test}
\title{Test the issue}
\usage{
test(vec)
}
\arguments{
\item{vec}{Test vector}
}
\description{
This function just calls into C++ to convert its argument to a std::array,
and then prints its elements. An error will be produced if \c vec doesn't
have three elements or can't be converted to double.
}
