% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{package}
\name{RcppArray-package}
\alias{RcppArray}
\alias{RcppArray-package}
\title{The 'RcppArray' package}
\description{
This package presents no user-facing R code, but instead provides linked
packages containing C++ code with the ability to convert easily between R
vector types and the \code{std::array} and \code{std::span} classes. See
the package README and/or included header \code{RcppArray.h} for further
details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonclayden/RcppArray}
  \item Report bugs at \url{https://github.com/jonclayden/RcppArray/issues}
}

}
\author{
\strong{Maintainer}: Jon Clayden \email{code@clayden.org} (\href{https://orcid.org/0000-0002-6608-0619}{ORCID})

Authors:
\itemize{
  \item Dirk Eddelbuettel
}

}
\keyword{internal}
